Move field into risc0-core crate
