+++ risc0/circuit/rv32im/Cargo.toml
@@ -13,6 +13,7 @@ harness = false
 
 [dependencies]
 anyhow = { version = "1.0", default-features = false }
+risc0-core = { workspace = true }
 risc0-zkp = { workspace = true }
 tracing = { version = "0.1", default-features = false, features = ["attributes"] }
