+++ risc0/circuit/rv32im/src/lib.rs
@@ -48,10 +48,8 @@ impl TapsProvider for CircuitImpl {
 
 #[cfg(test)]
 mod tests {
-    use risc0_zkp::{
-        adapter::{CircuitStep, CircuitStepContext, CircuitStepHandler},
-        field::baby_bear::BabyBearElem,
-    };
+    use risc0_core::field::baby_bear::BabyBearElem;
+    use risc0_zkp::adapter::{CircuitStep, CircuitStepContext, CircuitStepHandler};
 
     use crate::CircuitImpl;
 
@@ -95,12 +93,12 @@ mod tests {
 #[cfg(feature = "test")]
 pub mod testutil {
     use rand::{thread_rng, Rng};
+    use risc0_core::field::{
+        baby_bear::{BabyBearElem, BabyBearExtElem},
+        Elem, ExtElem,
+    };
     use risc0_zkp::{
         adapter::{CircuitInfo, TapsProvider},
-        field::{
-            baby_bear::{BabyBearElem, BabyBearExtElem},
-            Elem, ExtElem,
-        },
         hal::{Buffer, EvalCheck, Hal},
         taps::RegisterGroup,
         INV_RATE,
