+++ risc0/zkp/Cargo.toml
@@ -18,6 +18,7 @@ fil-rustacuda = { version = "0.1", optional = true }
 metal = { version = "0.24", optional = true }
 paste = "1.0"
 rand_core = "0.6"
+risc0-core = { workspace = true }
 risc0-sys = { workspace = true, optional = true }
 risc0-zeroio = { workspace = true }
 rustacuda_core = { version = "0.1", optional = true }
