+++ risc0/zkp/benches/ntt.rs
@@ -14,10 +14,8 @@
 
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use rand::thread_rng;
-use risc0_zkp::{
-    core::ntt::interpolate_ntt,
-    field::{baby_bear::BabyBearElem, Elem},
-};
+use risc0_core::field::{baby_bear::BabyBearElem, Elem};
+use risc0_zkp::core::ntt::interpolate_ntt;
 
 pub fn ntt(c: &mut Criterion) {
     let mut group = c.benchmark_group("interpolate_ntt");
