+++ risc0/zkvm/Cargo.toml
@@ -24,6 +24,7 @@ anyhow = { version = "1.0", default-features = false }
 bytemuck = "1.12"
 cfg-if = "1.0"
 risc0-circuit-rv32im = { workspace = true }
+risc0-core = { workspace = true }
 risc0-zeroio = { workspace = true }
 risc0-zkp = { workspace = true }
 risc0-zkvm-platform = { workspace = true }
